feat(settings): subscription section under settings; reasoning off by default
- Extract SubscriptionSummary (plan card, quick actions, credit bars) from
  SubscriptionView so both the standalone /subscription page and the new
  /settings/subscription section can share a single source of truth.
- Add "Subscription" as a section in SettingsView; sidebar profile
  dropdown still navigates to /subscription so both access paths work.
- Flip enableReasoning default from true to false. Extended thinking is
  opt-in going forward; existing user rows are untouched so users who
  explicitly turned it on keep their choice.
